refactor: move components to web for Tina types
